Reestructure project
