+++ Dockerfile
@@ -1,9 +1,11 @@
-FROM golang:1.21-alpine AS builder
+FROM golang:1.24-alpine AS builder
 
 WORKDIR /app
-COPY main.go .
-RUN go mod init ircbot && go mod tidy
-RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o ircbot main.go
+COPY go.mod ./
+COPY go.sum* ./
+RUN go mod download
+COPY . .
+RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o ircbot .
 
 FROM alpine:latest
 RUN apk --no-cache add ca-certificates
