+++ test_enhanced_bot_test.go
